用cQueue实现，发送达到4000每秒，但不稳定，应该是存取队列出问题

+++ main/include/msg.h
@@ -97,7 +97,7 @@ typedef struct
 {
     CAN_TX_MSGOBJ txObj;
     u8 channel;
-    u16 frames;
+    u8 frames;
     u8 data[1408];
 } CAN_CMD_MULTIFRAME;
